Tests passing with new APIs.

+++ src/ffi/linux/pcm.rs
@@ -346,7 +346,7 @@ pub(crate) unsafe fn state(pcm: *mut c_void) -> SndPcmState {
 pub(crate) unsafe fn readi<T>(
     pcm: *mut c_void,
     buffer: *mut T,
-    length: u8,
+    length: u16,
 ) -> Result<usize, i64> {
     ALSA.with(|alsa| {
         let alsa = if let Some(alsa) = alsa {
@@ -366,7 +366,7 @@ pub(crate) unsafe fn readi<T>(
 pub(crate) unsafe fn writei<T>(
     pcm: *mut c_void,
     buffer: *const T,
-    length: u8,
+    length: u16,
 ) -> Result<usize, i64> {
     ALSA.with(|alsa| {
         let alsa = if let Some(alsa) = alsa {
